+++ src/main/java/com/dj/repository/GameGenreRepository.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
  * Created by DJ on 11/23/16.
  */
 @Repository
-public interface GameGenreRepository extends JpaRepository<GameGenre, Long> {
+public interface GameGenreRepository extends JpaRepository<GameGenre, Integer> {
 	
 }
