+++ src/main/java/com/dj/repository/ScoreRepository.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
  * Created by DJ on 11/23/16.
  */
 @Repository
-public interface ScoreRepository extends JpaRepository<Score,Long> {
+public interface ScoreRepository extends JpaRepository<Score,Integer> {
 	
 }
