+++ src/main/java/com/dj/repository/WebsiteRepository.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
  * Created by DJ on 11/23/16.
  */
 @Repository
-public interface WebsiteRepository extends JpaRepository<Website,Long> {
+public interface WebsiteRepository extends JpaRepository<Website,Integer> {
 	
 }
